# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Service API v2
  description: |
    물품 배송 서비스 백엔드 API 명세서

    ## Architecture
    - Hexagonal Architecture (Ports & Adapters)
    - Spring Boot 4.0.1 / Java 21
    - Event-Driven Architecture

    ## Service Type
    - Product Delivery Service (물품 배송 서비스)
    - Courier-based delivery (택배사 연동)
    - Domestic shipping only (국내 배송)
    - Return/Exchange supported (반품/교환 지원)

    ## Domains
    1. Customer - 고객 관리
    2. Product - 상품 관리
    3. Seller - 판매자 관리
    4. Cart - 장바구니
    5. Order - 주문 관리
    6. Payment - 결제 관리
    7. Shipment - 배송 관리
    8. Return - 반품/교환 관리
    9. Review - 리뷰 관리
    10. Promotion - 쿠폰/프로모션
  version: 2.0.0
  contact:
    name: Delivery Service Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v2
    description: Local Development Server
  - url: https://api.delivery.com/api/v2
    description: Production Server
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: 이미 존재하는 이메일
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 인증 실패

  # ============================================
  # Customer API
  # ============================================
  /customers/me:
    get:
      tags: [Customers]
      summary: 내 정보 조회
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Customers]
      summary: 내 정보 수정
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
...
  /cart:
    get:
      tags: [Cart]
      summary: 장바구니 조회
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

    delete:
      tags: [Cart]
      summary: 장바구니 비우기
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 삭제 성공
  /cart/items:
    post:
      tags: [Cart]
      summary: 장바구니 상품 추가
      operationId: addToCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '201':
          description: 추가 성공
  /cart/items/{itemId}:
    patch:
      tags: [Cart]
      summary: 장바구니 수량 변경
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: 수정 성공

    delete:
      tags: [Cart]
      summary: 장바구니 상품 삭제
      operationId: removeFromCart
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 삭제 성공

  # ============================================
  # Order API
  # ============================================
  /orders:
    get:
      tags: [Orders]
      summary: 내 주문 목록 조회
      operationId: getMyOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

    post:
      tags: [Orders]
      summary: 주문 생성
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 주문 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: 주문 상세 조회
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: 주문 취소
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          description: 취소 성공
        '409':
          $ref: '#/components/responses/Conflict'
...
  /sellers/me/orders/{orderId}/confirm:
    post:
      tags: [Seller Admin]
      summary: 주문 확정
      operationId: confirmOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 확정 성공
  /sellers/me/orders/{orderId}/prepare:
    post:
      tags: [Seller Admin]
      summary: 상품 준비 시작
      operationId: startPreparing
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 준비 시작 성공
  /sellers/me/orders/{orderId}/ship:
    post:
      tags: [Seller Admin]
      summary: 출고 완료 (운송장 등록)
      operationId: shipOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipOrderRequest'
      responses:
        '200':
          description: 출고 성공
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: 입력값 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 상태 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - PENDING
        - PAID
        - CONFIRMED
        - PREPARING
        - SHIPPED
        - IN_TRANSIT
        - OUT_FOR_DELIVERY
        - DELIVERED
        - CANCELLED
        - RETURN_REQUESTED
        - RETURNED
    ShipmentStatus:
      type: string
      enum:
        - PENDING
        - READY_TO_SHIP
        - PICKED_UP
        - IN_TRANSIT
        - OUT_FOR_DELIVERY
        - DELIVERED
        - FAILED
        - RETURNED
    ShippingCarrier:
      type: string
      enum:
        - CJ_LOGISTICS
        - HANJIN
        - LOTTE
        - LOGEN
        - POST_OFFICE
        - EPOST
        - GS_POSTBOX
        - CU_POST
        - OTHER
...
    RegisterRequest:
      type: object
      required: [email, password, name, phone]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        profileImageUrl:
          type: string
...
    AddToCartRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        variantId:
          type: string
        quantity:
          type: integer
          minimum: 1
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
    CreateOrderRequest:
      type: object
      required: [sellerId, items, shippingAddress]
      properties:
        sellerId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddressRequest'
        orderMemo:
          type: string
        couponId:
          type: string
    OrderItemRequest:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        variantName:
          type: string
        sku:
          type: string
        optionValues:
          type: object
          additionalProperties:
            type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
    ShippingAddressRequest:
      type: object
      required: [recipientName, phoneNumber, postalCode, address1]
      properties:
        recipientName:
          type: string
        phoneNumber:
          type: string
        postalCode:
          type: string
        address1:
          type: string
        address2:
          type: string
        deliveryNote:
          type: string
    CancelOrderRequest:
      type: object
      properties:
        reason:
          type: string
...
    ShipOrderRequest:
      type: object
      required: [carrier, trackingNumber]
      properties:
        carrier:
          $ref: '#/components/schemas/ShippingCarrier'
        trackingNumber:
          type: string
...
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        profileImageUrl:
          type: string
        status:
          type: string
        pointBalance:
          type: number
        createdAt:
          type: string
          format: date-time
...
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemResponse'
        totalAmount:
          type: number
        totalItems:
          type: integer
    CartItemResponse:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        variantName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        thumbnailUrl:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        stock:
          type: integer
        isAvailable:
          type: boolean
    OrderResponse:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerId:
          type: string
        sellerId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotalAmount:
          type: number
        shippingFee:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddressResponse'
        createdAt:
          type: string
          format: date-time
    OrderItemResponse:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        variantName:
          type: string
        sku:
          type: string
        optionValues:
          type: object
          additionalProperties:
            type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    ShippingAddressResponse:
      type: object
      properties:
        recipientName:
          type: string
        phoneNumber:
          type: string
        postalCode:
          type: string
        address1:
          type: string
        address2:
          type: string
        deliveryNote:
          type: string
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OrderResponse'
        - type: object
          properties:
            sellerName:
              type: string
            orderMemo:
              type: string
            shipment:
              $ref: '#/components/schemas/ShipmentSummaryResponse'
            paidAt:
              type: string
              format: date-time
            confirmedAt:
              type: string
              format: date-time
            shippedAt:
              type: string
              format: date-time
            deliveredAt:
              type: string
              format: date-time
    OrderListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummaryResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
    OrderSummaryResponse:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
        itemCount:
          type: integer
        thumbnailUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
...
    ShipmentSummaryResponse:
      type: object
      properties:
        id:
          type: string
        carrier:
          $ref: '#/components/schemas/ShippingCarrier'
        carrierName:
          type: string
        trackingNumber:
          type: string
        trackingUrl:
          type: string
        status:
          $ref: '#/components/schemas/ShipmentStatus'
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time